requirements.txt
python-dotenv
streamlit
sentence-transformers
langchain-community
langchain
langchain-huggingface
langchain-text-splitters
unstructured
unstructured[pdf]
langchain-unstructured
langchain-chroma
langchain-groq
unstructured-inference